Add zscore
